Refactor SearchResultForm and ProcessForm to align available actions
