+++ Kitodo/src/main/webapp/WEB-INF/templates/includes/processes/statistics.xhtml
@@ -1,0 +1,87 @@
+<!--
+ *
+ * (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
+ *
+ * This file is part of the Kitodo project.
+ *
+ * It is licensed under GNU General Public License version 3 or later.
+ *
+ * For the full copyright and license information, please read the
+ * GPL3-License.txt file that was distributed with this source code.
+ *
+-->
+
+<ui:composition
+        xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+        xmlns:h="http://xmlns.jcp.org/jsf/html"
+        xmlns:p="http://primefaces.org/ui">
+
+    <!--@elvariable id="StatisticView" type="org.kitodo.production.forms.ProcessListBaseView"-->
+    <h:panelGroup id="statisticPanel">
+        <p:pieChart rendered="#{StatisticView.showPieModel()}"
+                    model="#{StatisticView.pieModel}"/>
+        <p:dataTable  id="statisticResult"
+                      style="width: 50%;"
+                      rendered="#{StatisticView.showPieModel()}"
+                      value="#{StatisticView.statisticResult}"
+                      var="result">
+            <p:column headerText="#{msgs.task}"
+                      width="50">
+                <h:outputText title="#{result.key}"
+                              value="#{result.key}"/>
+            </p:column>
+            <p:column headerText="#{msgs.amount}"
+                      width="50">
+                <h:outputText title="#{result.value}"
+                              value="#{result.value}"/>
+            </p:column>
+        </p:dataTable>
+
+        <p:barChart rendered="#{StatisticView.showBarModel()}"
+                    model="#{StatisticView.stackedBarModel}"
+                    style="width: 100%;"/>
+
+        <p:dataTable id="processMetadataTable"
+                     style="width: 50%;"
+                     rendered="#{StatisticView.showProcessMetadataStatisticTable()}"
+                     value="#{StatisticView.processMetadataStatistics}"
+                     var="processStatistic">
+            <p:column headerText="#{msgs.title}"
+                      width="25">
+                <h:outputText title="#{processStatistic.processTitle}"
+                              value="#{processStatistic.processTitle}"/>
+            </p:column>
+            <p:column headerText="#{msgs.images}"
+                      width="10">
+                <h:outputText title="#{processStatistic.numberOfImages}"
+                              value="#{processStatistic.numberOfImages}"/>
+            </p:column>
+            <p:column headerText="#{msgs.relativeAmount}"
+                      width="10">
+                <p:progressBar value="#{StatisticView.getRelativeImageAmount(processStatistic.numberOfImages)}"
+                               displayOnly="true"/>
+            </p:column>
+            <p:column headerText="#{msgs.structuralElement}"
+                      width="10">
+                <h:outputText title="#{processStatistic.numberOfStructuralElements}"
+                              value="#{processStatistic.numberOfStructuralElements}"/>
+            </p:column>
+            <p:column headerText="#{msgs.relativeAmount}"
+                      width="10">
+                <p:progressBar value="#{StatisticView.getRelativeStructuralElementAmount(processStatistic.numberOfStructuralElements)}"
+                               displayOnly="true"/>
+            </p:column>
+            <p:column headerText="#{msgs.metadata}"
+                      width="10">
+                <h:outputText title="#{processStatistic.numberOfMetadata}"
+                              value="#{processStatistic.numberOfMetadata}"/>
+            </p:column>
+            <p:column headerText="#{msgs.relativeAmount}"
+                      width="10">
+                <p:progressBar value="#{StatisticView.getRelativeMetadataAmount(processStatistic.numberOfMetadata)}"
+                               displayOnly="true"/>
+            </p:column>
+        </p:dataTable>
+    </h:panelGroup>
+
+</ui:composition>
