+++ Kitodo/src/main/webapp/pages/processes.xhtml
@@ -104,7 +104,9 @@
     <ui:define name="dialog">
         <ui:include src="/WEB-INF/templates/includes/processes/batchCreatePopup.xhtml"/>
         <ui:include src="/WEB-INF/templates/includes/processes/batchRenamePopup.xhtml"/>
-        <ui:include src="/WEB-INF/templates/includes/processes/deleteChildrenDialog.xhtml"/>
         <ui:include src="/WEB-INF/templates/includes/processes/executeScriptSelectedPopup.xhtml"/>
+        <ui:include src="/WEB-INF/templates/includes/processes/deleteChildrenDialog.xhtml">
+            <ui:param name="ProcessListView" value="#{ProcessForm}"/>
+        </ui:include>
     </ui:define>
 </ui:composition>
